authentication system & size limit

+++ src/db/fileDatabase.ts
@@ -24,8 +24,8 @@ pool.connect();
 console.log("Connected to database");
 
 export default class FileDatabase {
-  static async listFiles(): Promise<FileData[]> {
-    return await pool.query("select * from files where deletedat is null;")
+  static async listFiles(userID: string): Promise<FileData[]> {
+    return await pool.query("select * from files where deletedat is null and author::text = $1;", [userID])
       .then(({ rows }) => filesTransformer(rows))
       .catch((err) => {
         console.log(err);
@@ -33,8 +33,8 @@ export default class FileDatabase {
       })
   }
 
-  static async listArchivedFiles(): Promise<FileData[]> {
-    return await pool.query("select * from files where deletedat is not null;")
+  static async listArchivedFiles(userID: string): Promise<FileData[]> {
+    return await pool.query("select * from files where deletedat is not null and author::text = $1;", [userID])
       .then(({ rows }) => filesTransformer(rows))
       .catch((err) => {
         console.log(err);
@@ -42,16 +42,12 @@ export default class FileDatabase {
       })
   }
 
-  static async enterFile(
-    fileID: string,
-    name: string,
-    author: string = "anon",
-  ): Promise<FileData> {
+  static async enterFile(fileID: string, name: string, userID: string): Promise<FileData> {
     if (!validate(fileID)) {
       throw new InternalServerException("generating a new file ID");
     }
 
-    return await pool.query("insert into files (fileid, name, author) values ($1::uuid, $2, $3) returning *;", [fileID, name, author])
+    return await pool.query("insert into files (fileid, name, author) values ($1::uuid, $2, $3::uuid) returning *;", [fileID, name, userID])
       .then(async ({ rows }) => {
         if (rows.length === 0) {
           throw new InternalServerException("entering a new file");
@@ -64,26 +60,23 @@ export default class FileDatabase {
       })
   }
 
-  static async findFile(fileID: string): Promise<FileData> {
-    const json = await this.getJSON(fileID);
+  static async findFile(fileID: string, userID: string): Promise<FileData> {
+    const json = await this.getJSON(fileID, userID);
 
     const file = fileTransformer(json);
 
     return file;
   }
 
-  static async renameFile(
-    fileID: string,
-    name: string,
-  ): Promise<FileData> {
-    const json = await this.getJSON(fileID);
+  static async renameFile(fileID: string, name: string, userID: string): Promise<FileData> {
+    const json = await this.getJSON(fileID, userID);
 
     const oldFile = fileTransformer(json);
     if (oldFile.deletedat) {
       throw new DeletedException("File", fileID);
     }
 
-    return await pool.query("update files set name = $1 where fileID::text = $2 returning *;", [name, fileID])
+    return await pool.query("update files set name = $1 where fileID::text = $2 and author::text = $3 returning *;", [name, fileID, userID])
       .then(async ({ rows }) => {
         if (rows.length === 0) {
           throw new InternalServerException("renaming the file");
@@ -98,8 +91,8 @@ export default class FileDatabase {
       })
   }
 
-  static async archiveFile(fileID: string): Promise<FileData> {
-    const json = await this.getJSON(fileID);
+  static async archiveFile(fileID: string, userID: string): Promise<FileData> {
+    const json = await this.getJSON(fileID, userID);
 
     const oldFile = fileTransformer(json);
 
@@ -107,7 +100,7 @@ export default class FileDatabase {
       throw new DeletedException("File", fileID);
     }
 
-    return await pool.query("update files set deletedat = CURRENT_TIMESTAMP where fileID::text = $1 returning *;", [fileID])
+    return await pool.query("update files set deletedat = CURRENT_TIMESTAMP where fileID::text = $1 and author::text = $2 returning *;", [fileID, userID])
       .then(async ({ rows }) => {
         if (rows.length === 0) {
           throw new InternalServerException("deleting the file");
@@ -121,8 +114,8 @@ export default class FileDatabase {
       })
   }
 
-  static async restoreFile(fileID: string): Promise<FileData> {
-    const json = await this.getJSON(fileID);
+  static async restoreFile(fileID: string, userID: string): Promise<FileData> {
+    const json = await this.getJSON(fileID, userID);
 
     const oldFile = fileTransformer(json);
 
@@ -130,7 +123,7 @@ export default class FileDatabase {
       throw new DeletedException("File", fileID, false);
     }
 
-    return await pool.query("update files set deletedat = null where fileID::text = $1 returning *;", [fileID])
+    return await pool.query("update files set deletedat = null where fileID::text = $1 and author::text = $2 returning *;", [fileID, userID])
       .then(async ({ rows }) => {
         if (rows.length === 0) {
           throw new InternalServerException("restoring the file");
@@ -144,12 +137,12 @@ export default class FileDatabase {
       })
   }
 
-  static async deleteFile(fileID: string): Promise<FileData> {
+  static async deleteFile(fileID: string, userID: string): Promise<FileData> {
     if (!validate(fileID)) {
       throw new DataNotFoundException("File", fileID);
     }
 
-    const json = await this.getJSON(fileID);
+    const json = await this.getJSON(fileID, userID);
 
     const file = fileTransformer(json);
 
@@ -157,7 +150,7 @@ export default class FileDatabase {
       throw new DeletedException("File", fileID, false);
     }
 
-    return await pool.query("delete from files where fileID::text = $1;", [fileID])
+    return await pool.query("delete from files where fileID::text = $1 and author::text = $2;", [fileID, userID])
       .then(async () => {
         console.log(`File ${fileID} deleted from the database permanently.`);
         return file;
@@ -168,8 +161,8 @@ export default class FileDatabase {
       })
   }
 
-  private static async getJSON(fileID: string): Promise<JSON> {
-    return await pool.query("select * from files where fileID::text = $1;", [fileID])
+  private static async getJSON(fileID: string, userID: string): Promise<JSON> {
+    return await pool.query("select * from files where fileID::text = $1 and author::text = $2;", [fileID, userID])
       .then(async ({ rows }) => {
         if (rows.length === 0) {
           console.log(`File ${fileID} does not exist in database.`);
